+++ sabka-bazar/src/components/cart/cart.scss
@@ -84,6 +84,7 @@
             overflow: hidden;
             h2 {
                 margin-left: 5%;
+                margin-top: 5%;
                 font-weight: bold;
             }
