+++ sabka-bazar/src/components/home/home.scss
