+++ sabka-bazar/src/components/model/model.scss
