+++ sabka-bazar/src/components/plp/plp.scss
