+++ sabka-bazar/src/containers/home/index.js
@@ -1,31 +1,8 @@
-import React, { useEffect } from "react";
 import Home from '../../components/home';
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { requestBannerData, requestCategoriesData } from "./actions";
 import { requestAddProduct, requestMinusProduct } from "../plp/actions";
-
-
-function HomeContainer(props) {
-
-    const { banners, categories, cartProducts } = props;
-
-    useEffect(() => {
-        props.requestBannerData();
-        props.requestCategoriesData();
-    }, []);
-
-    function plusProduct(product) {
-        props.requestAddProduct(product);
-    }
-
-    function minusProduct(product) {
-        props.requestMinusProduct(product);
-    }
-
-
-    return <Home banners={banners} categories={categories} cartProducts={cartProducts} plusProduct={plusProduct} minusProduct={minusProduct} />;
-}
 
 const mapStateToProps = state => ({
     banners: state.home.banners,
@@ -35,4 +12,4 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = dispatch => bindActionCreators({ requestBannerData, requestCategoriesData, requestAddProduct, requestMinusProduct }, dispatch);
 
-export default connect(mapStateToProps, mapDispatchToProps)(HomeContainer);
+export default connect(mapStateToProps, mapDispatchToProps)(Home);
