+++ sabka-bazar/src/style/common/style.scss
@@ -3,6 +3,7 @@
     font-family: 'Dosis','Arial';
     font-size: 14px;
     box-sizing: border-box;
+    margin: 0px;
     }
  
     a {
@@ -195,7 +196,7 @@ header {
         opacity: 0.7;
         text-align: center;
         margin-left: 60px;
-        margin-top: 25px;
+        margin-top: 25%;
         img {
             width: 25px;
             height: 25px;
